gwak/deploy/pyproject.toml
[tool.poetry]
name = "deploy"
version = "0.1.0"
description = ""
authors = ["Andy Chen <andy.c.80297@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"

# ml4gw applications (local) 
ml4gw = {path = "../../ml4gw", develop = true}
ml4gw-hermes = {path = "../../hermes", develop = true}

spython = "0.2.14"
nvidia-cudnn-cu11 = "8.9.6.50"
tensorrt = "8.5.2.2"

# workaround for: https://github.com/python-poetry/poetry-plugin-export/issues/183
urllib3 = ">=1.25.4,<1.27"
jsonargparse = "^4.36.0"
onnx = "^1.17.0"
psutil = "^6.1.1"
h5py = "^3.12.1"
tqdm = "^4.67.1"

[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
